apps/asset-service/requirements.txt
# TB ERP - Asset Service Dependencies
# FastAPI and async support
# FastAPI and async support
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
gunicorn>=23.0.0

# Database
sqlalchemy[asyncio]>=2.0.35
asyncpg>=0.30.0
alembic>=1.14.0

# Validation and Settings
pydantic>=2.9.0
pydantic-settings>=2.6.0


# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Redis
redis==5.0.1

# HTTP Client
httpx==0.26.0

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0

# Linting
ruff==0.1.8
